moving all the non-db stuff under self-serve
