+++ app/blueprints/dev/routes.py
@@ -4,13 +4,13 @@ from flask import redirect
 from flask import request
 from flask import url_for
 
-from app.data.data_access import clear_all_responses
-from app.data.data_access import get_all_components
-from app.data.data_access import get_all_pages
-from app.data.data_access import get_all_sections
-from app.data.data_access import get_responses
-from app.data.data_access import get_saved_forms
-from app.data.data_access import save_response
+from app.blueprints.self_serve.data.data_access import clear_all_responses
+from app.blueprints.self_serve.data.data_access import get_all_components
+from app.blueprints.self_serve.data.data_access import get_all_pages
+from app.blueprints.self_serve.data.data_access import get_all_sections
+from app.blueprints.self_serve.data.data_access import get_responses
+from app.blueprints.self_serve.data.data_access import get_saved_forms
+from app.blueprints.self_serve.data.data_access import save_response
 
 dev_bp = Blueprint(
     "dev_bp",
